Created component to display state for a sample analysis
